Refactor operator strategy
Split code related to strategy to separate
files
